Estoque Atualizado/requirements.txt
# Dependências do projeto
# Instale com: pip install -r requirements.txt

flask==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
